refactor CMS to use context api
