+++ cms/app/edit/[id]/page.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 import { useParams, useRouter } from "next/navigation";
 import Link from "next/link";
 import { Scheherazade_New } from 'next/font/google';
-import DoaPreview from "../../component/doa-preview";
+import DoaPreview from "../../component/DoaPreview";
 import {DuaType} from "../../../../config/types";
 
 interface Subcategory {
